Import funguloids-linux-1.05.tar.bz2 (version 1.05)
These files have been unpacked:

bin/bootstrap.mpk
bin/funguloids.mpk

These files have been converted:

bin/funguloids/whiteglow2.dds
bin/funguloids/bg2.dds
bin/funguloids/bg4.dds
bin/funguloids/bg8.dds
bin/funguloids/bg1.dds
bin/funguloids/fieryglow.dds
bin/funguloids/bg5.dds
bin/funguloids/bg3.dds
bin/funguloids/whiteglow.dds
bin/funguloids/bg7.dds
bin/funguloids/bg6.dds
bin/funguloids/logo.dds
bin/funguloids/Plane.mesh
bin/funguloids/Drop.mesh
bin/funguloids/Tentacle.mesh
bin/funguloids/Asteroid1.mesh
bin/funguloids/Base.mesh
bin/funguloids/Whirler.mesh
bin/funguloids/Mushroom.mesh
bin/funguloids/Asteroid2.mesh
bin/funguloids/Asteroid3.mesh
bin/funguloids/WormBall.mesh
bin/funguloids/Base-Armature.skeleton

These files have been deleted:

bin/bootstrap.mpk
bin/funguloids.mpk
bin/funguloids/whiteglow2.dds
bin/funguloids/bg2.dds
bin/funguloids/bg4.dds
bin/funguloids/bg8.dds
bin/funguloids/bg1.dds
bin/funguloids/fieryglow.dds
bin/funguloids/bg5.dds
bin/funguloids/bg3.dds
bin/funguloids/whiteglow.dds
bin/funguloids/bg7.dds
bin/funguloids/bg6.dds
bin/funguloids/logo.dds
bin/funguloids/Plane.mesh
bin/funguloids/Drop.mesh
bin/funguloids/Tentacle.mesh
bin/funguloids/Asteroid1.mesh
bin/funguloids/Base.mesh
bin/funguloids/Whirler.mesh
bin/funguloids/Mushroom.mesh
bin/funguloids/Asteroid2.mesh
bin/funguloids/Asteroid3.mesh
bin/funguloids/WormBall.mesh
bin/funguloids/Base-Armature.skeleton
include/SimpleIni
bin/bootstrap/VeraIt.ttf
bin/bootstrap/Vera.ttf
bin/bootstrap/VeraBd.ttf
bin/bootstrap/VeraBI.ttf
obj/empty
obj

+++ src/scriptsystem.cpp
@@ -153,7 +153,7 @@ static int CreateMushroom(lua_State *L) {
 
 	// Get the params
 	const char *name = lua_tostring(L, 1);
-	int color = lua_tonumber(L, 2);
+	int color = (int)lua_tonumber(L, 2);
 	Real xpos = lua_tonumber(L, 3);
 	Real ypos = lua_tonumber(L, 4);
 
@@ -220,7 +220,7 @@ static int CreateBase(lua_State *L) {
 
 	// Get the params
 	const char *name = lua_tostring(L, 1);
-	int color = lua_tonumber(L, 2);
+	int color = (int)lua_tonumber(L, 2);
 	Real xpos = lua_tonumber(L, 3);
 	Real ypos = lua_tonumber(L, 4);
 	const char *level;
